RavenDB-26628 Refactor
