+++ src/Raven.AiAppliance.Web/.prettierrc
@@ -1,11 +1,11 @@
 {
-  "endOfLine": "lf",
-  "semi": true,
-  "singleQuote": false,
-  "tabWidth": 2,
-  "trailingComma": "all",
-  "printWidth": 80,
-  "plugins": ["prettier-plugin-tailwindcss"],
-  "tailwindStylesheet": "src/index.css",
-  "tailwindFunctions": ["cn", "cva"]
+    "endOfLine": "lf",
+    "semi": true,
+    "singleQuote": false,
+    "tabWidth": 4,
+    "trailingComma": "all",
+    "printWidth": 120,
+    "plugins": ["prettier-plugin-tailwindcss"],
+    "tailwindStylesheet": "src/index.css",
+    "tailwindFunctions": ["cn", "cva"]
 }
