+++ src/Raven.AiAppliance.Web/pnpm-workspace.yaml
@@ -1,2 +1,2 @@
 allowBuilds:
-  msw: false
+    msw: false
