+++ src/Raven.AiAppliance.Web/src/components/auth/auth-context.ts
@@ -1,20 +1,20 @@
 import { createContext, useContext } from "react";
-import type { LoginRequest } from "@/api/authApi";
+import type { LoginRequest } from "@/api/auth-service";
 
 export type AuthContextValue = {
-  isAuthenticated: boolean;
-  isLoading: boolean;
-  login: (request: LoginRequest) => Promise<boolean>;
+    isAuthenticated: boolean;
+    isLoading: boolean;
+    login: (request: LoginRequest) => Promise<boolean>;
 };
 
 export const AuthContext = createContext<AuthContextValue | null>(null);
 
 export function useAuth() {
-  const context = useContext(AuthContext);
+    const context = useContext(AuthContext);
 
-  if (!context) {
-    throw new Error("useAuth must be used within AuthProvider.");
-  }
+    if (!context) {
+        throw new Error("useAuth must be used within AuthProvider.");
+    }
 
-  return context;
+    return context;
 }
