--- src/Raven.AiAppliance.Web/src/components/ui/separator.tsx
@@ -1,28 +1,0 @@
-"use client";
-
-import * as React from "react";
-import { Separator as SeparatorPrimitive } from "radix-ui";
-
-import { cn } from "@/lib/utils";
-
-function Separator({
-  className,
-  orientation = "horizontal",
-  decorative = true,
-  ...props
-}: React.ComponentProps<typeof SeparatorPrimitive.Root>) {
-  return (
-    <SeparatorPrimitive.Root
-      data-slot="separator"
-      decorative={decorative}
-      orientation={orientation}
-      className={cn(
-        "shrink-0 bg-border data-horizontal:h-px data-horizontal:w-full data-vertical:w-px data-vertical:self-stretch",
-        className,
-      )}
-      {...props}
-    />
-  );
-}
-
-export { Separator };
