+++ src/Raven.AiAppliance.Web/src/main.tsx
@@ -1,23 +1,23 @@
 import "./index.css";
+import { QueryClientProvider } from "@tanstack/react-query";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { ThemeProvider } from "@/components/theme-provider.tsx";
 import { AuthProvider } from "@/components/auth/auth-provider";
-import { RouterProvider } from "react-router";
-import { router } from "@/routes.tsx";
-import { QueryClientProvider } from "@tanstack/react-query";
+import { ThemeProvider } from "@/components/shadcn/theme-provider";
 import { queryClient } from "@/lib/query-client";
-import { Toaster } from "@/components/ui/sonner";
+import { router } from "@/routes";
+import { RouterProvider } from "react-router";
+import { Toaster } from "@/components/shadcn/ui/sonner";
 
 createRoot(document.getElementById("root")!).render(
-  <StrictMode>
-    <ThemeProvider>
-      <QueryClientProvider client={queryClient}>
-        <AuthProvider>
-          <RouterProvider router={router} />
-          <Toaster />
-        </AuthProvider>
-      </QueryClientProvider>
-    </ThemeProvider>
-  </StrictMode>,
+    <StrictMode>
+        <ThemeProvider>
+            <QueryClientProvider client={queryClient}>
+                <AuthProvider>
+                    <RouterProvider router={router} />
+                    <Toaster />
+                </AuthProvider>
+            </QueryClientProvider>
+        </ThemeProvider>
+    </StrictMode>,
 );
